Fikset att man ikke teller antall perioder men antall oppdrag
